+++ src/pages/Home.css
@@ -1,0 +1,8 @@
+section{
+    padding: 1rem 2rem;
+}
+
+.loading-container{
+    display: flex;
+    justify-content: center;
+}
