--- src/pages/Home.module.css
@@ -1,3 +1,0 @@
-section{
-    padding: 1rem 2rem;
-}
